+++ tools/junit2slack/go.mod
@@ -1,16 +1,16 @@
-module junit2slack
+module github.com/stackrox/stackrox/tools/junit2slack
 
 go 1.18
 
 require (
 	github.com/GoogleCloudPlatform/testgrid v0.0.147
 	github.com/slack-go/slack v0.11.3
-	github.com/stretchr/testify v1.7.0
+	github.com/stretchr/testify v1.8.0
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/gorilla/websocket v1.4.2 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
