+++ tools/junit2slack/testdata/sample.xml
